feat: add copy file

+++ src/fs/create.js
@@ -1,10 +1,15 @@
 import { writeFile } from "node:fs/promises";
 import { getPathToFile } from "../utils/pathToFile.js";
+import { checkFileExist } from "../utils/checkFileExist.js";
 
 const create = async () => {
+  const fileUrl = getPathToFile(import.meta.url, "files", "fresh.txt");
+  const data = "I am fresh and young";
+
   try {
-    const fileUrl = getPathToFile(import.meta.url, "files", "fresh.txt");
-    const data = "I am fresh and young";
+    if (await checkFileExist(fileUrl)) {
+      throw new Error("FS operation failed");
+    }
     await writeFile(fileUrl, data);
   } catch (err) {
     console.error(err);
